requirements.txt
snowflake-connector-python>=3.12.3
pandas